線が引けないぜ☆（＾～＾）

+++ piston-pixel-art/src/res/settings.rs
@@ -14,7 +14,7 @@ pub struct Settings {
     // 画像ファイルの縦幅
     pub height: u32,
     // 拡大率
-    pub zoom: u8,
+    pub zoom: f64,
     // クライアント領域の左上からのオフセット
     pub canvas_margin_x: f64,
     pub canvas_margin_y: f64,
